tests/wpt/web-platform-tests/trusted-types/HTMLAreaElement-href.tentative.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="support/helper.sub.js"></script>
<body>
  <map>
    <script>
    //helper function for the tests
    function testHref(str, url) {
      var area = document.createElement('area');
      area.href = url;
      area.alt = "Area";
      assert_equals(area.href, str);
    }

    test(t => {
      testHref(URLS.safe, TrustedURL.create(URLS.safe));
    }, "area.href = URLS.safe, TrustedURL.create");

    test(t => {
      testHref(URLS.safe, TrustedURL.unsafelyCreate(URLS.safe));
    }, "area.href = URLS.safe, TrustedURL.unsafelyCreate");
    </script>
  </map>
</body>
